requirements.txt
mcp
aiofiles
Pillow
requests
